requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
bitarray==2.9.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cytoolz==0.12.3
Django==5.0.4
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
frozenlist==1.4.1
gunicorn==21.2.0
hexbytes==0.3.1
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
lru-dict==1.2.0
multidict==6.0.5
packaging==24.0
parsimonious==0.10.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==15.1.0
redis==5.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
web3==6.17.0
websockets==12.0
yarl==1.9.4
